Add redis input as plugin
